feat: move state to html and implement accessible status
Based on the logic provided by @evert for checking what to do next with each flask.
The references were tweaked so the state is kept in the HTML.

Added some better semantic markup to the content and specific ARIA roles so the game can be played with keyboard and a screen reader.